requirements.txt
matplotlib==3.4.1
pandas==1.2.4
scikit_image==0.18.1
streamlit==0.80.0
firebase_admin==4.5.3
tensorflow==2.3.0
streamlit_drawable_canvas==0.6.0
numpy==1.20.2
Pyrebase==3.0.27
appJar==0.94.0
ipython==7.22.0
JSAnimation==0.1
Pillow==8.2.0
psutil==5.8.0
pyserial==3.5
scikit_learn==0.24.1
seaborn==0.11.1
skimage==0.0
torch==1.8.1
torchvision==0.9.1
